% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicUtils.R
\name{fortify.df}
\alias{fortify.df}
\title{Create a fortified dataframe.}
\usage{
fortify.df(dfList, fInterval, by, names)
}
\arguments{
\item{dfList}{a list of dataframes with a dates column in \code{\link[base]{POSIXct}} format an another column with any value type.}

\item{fInterval}{a 2-item vector with 'start' and 'end' POSIXct dates respectively. All the values in the given interval will be explicitdly
included in the resulting dataframe with NAs. By default the limit dates are extracted from the data, in this case 'start' and end'
its the minimun and maximun dates for \strong{all} the dataframes given in \code{dfList}.}

\item{by}{a character string, containing the data increment in terms of "sec", "min", "hour", "day", "DSTday", "week", "month", "quarter" or "year".
This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".
See \code{\link[base]{seq.POSIXt}} for more info. The increment provided must equivalent to the real data-time increment.}

\item{names}{a vector of names for the dataframe \strong{values}.}
}
\value{
A fortified dataframe, i.e with all the lost values explicited. The resulting time index its perfectly
defined by \code{seq(start,end,by)}, so the columns of the dataframe created with this function will have the same index,
i.e the silent lost values of the inputs dataframes (the columns in the new one) will be marked with NAs in the right positions.
}
\description{
Create a fortified dataframe.
}
\examples{
fortify.df(dfList=myDfList, by="30 min", names = c("A","B","Z"))
fortify.df(df,c(as.POSIXct("1/01/2019"), as.POSIXct("1/01/2020")),
           "hour", names = c("A","B","Z"))
}
\author{
JV Yago
}
